refactor web client functions
