+++ csv/pokemonkanto.csv
@@ -15,7 +15,7 @@ kakuna,kanto
 beedrill,kanto
 pidgey,kanto
 pidgeotto,kanto
-pidgeo,kanto
+pidgeot,kanto
 rattata,kanto
 raticate,kanto
 spearow,kanto
